requirements.txt
annotated-types==0.7.0
blinker==1.8.2
cachetools==5.4.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
Flask==3.0.3
google-ai-generativelanguage==0.6.6
google-api-core==2.19.1
google-api-python-client==2.137.0
google-auth==2.32.0
google-auth-httplib2==0.2.0
google-generativeai==0.7.2
googleapis-common-protos==1.63.2
grpcio==1.64.1
grpcio-status==1.62.2
httplib2==0.22.0
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
proto-plus==1.24.0
protobuf==4.25.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
pydantic==2.8.2
pydantic_core==2.20.1
pyparsing==3.1.2
python-dotenv==1.0.1
requests==2.32.2
rsa==4.9
tqdm==4.66.4
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.1
waitress==3.0.0
Werkzeug==3.0.3
